+++ spec/resources/acl_spec.rb
@@ -1,10 +1,12 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 describe 'adding access control lists' do
   let(:chef_run) do
     ChefSpec::SoloRunner.new(
       platform: 'centos', version: '8', step_into: %w(bind_config bind_acl)
-    ).converge('bind_test::spec_acl')
+    ).converge('test::spec_acl')
   end
 
   include_context 'version_stub'