+++ spec/resources/service_spec.rb
@@ -1,10 +1,12 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 describe 'basic recipe on centos 8' do
   let(:chef_run) do
     ChefSpec::SoloRunner.new(
       platform: 'centos', version: '8', step_into: ['bind_service']
-    ).converge('bind_test::spec_basic')
+    ).converge('test::spec_basic')
   end
 
   include_context 'version_stub'
@@ -40,7 +42,7 @@ describe 'chroot recipe on ubuntu 18.04' do
   let(:chef_run) do
     ChefSpec::SoloRunner.new(
       platform: 'ubuntu', version: '18.04', step_into: ['bind_service']
-    ).converge('bind_test::spec_chroot')
+    ).converge('test::spec_chroot')
   end
 
   include_context 'version_stub'
@@ -116,7 +118,7 @@ describe 'chroot recipe on centos 8' do
   let(:chef_run) do
     ChefSpec::SoloRunner.new(
       platform: 'centos', version: '8', step_into: ['bind_service']
-    ).converge('bind_test::spec_chroot')
+    ).converge('test::spec_chroot')
   end
 
   include_context 'version_stub'
@@ -166,7 +168,7 @@ describe 'overridden defaults on centos 8' do
   let(:chef_run) do
     ChefSpec::SoloRunner.new(
       platform: 'centos', version: '8', step_into: ['bind_service']
-    ).converge('bind_test::spec_overridden')
+    ).converge('test::spec_overridden')
   end
 
   it 'creates configuration directories' do
@@ -185,7 +187,7 @@ describe 'basic recipe on ubuntu 18.04' do
   let(:chef_run) do
     ChefSpec::SoloRunner.new(
       platform: 'ubuntu', version: '18.04', step_into: ['bind_service']
-    ).converge('bind_test::spec_basic')
+    ).converge('test::spec_basic')
   end
 
   include_context 'version_stub'