+++ test/cookbooks/test/metadata.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+name 'test'
+version '1.0.0'
+maintainer 'Sous Chefs'
+maintainer_email 'help@sous-chefs.org'
+
+depends 'bind'