+++ test/cookbooks/test/recipes/disable_resolved.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+service 'systemd-resolved' do
+  action :stop
+end