+++ test/cookbooks/test/recipes/spec_basic.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+bind_service 'default' do
+  action [:create, :start]
+end
+
+bind_config 'default'