+++ test/fixtures/cookbooks/bind_test/metadata.rb
@@ -1,6 +1,0 @@
-name 'bind_test'
-version '1.0.0'
-maintainer 'David Bruce'
-maintainer_email 'djb@ragnarok.net'
-
-depends 'bind'